+++ library/core/src/main/java/com/google/android/exoplayer2/upstream/DefaultDataSourceFactory.java
@@ -59,7 +59,7 @@ public final class DefaultDataSourceFactory implements Factory {
    */
   public DefaultDataSourceFactory(
       Context context, String userAgent, @Nullable TransferListener listener) {
-    this(context, listener, new DefaultHttpDataSourceFactory(userAgent, listener));
+    this(context, listener, new DefaultHttpDataSource.Factory().setUserAgent(userAgent));
   }
 
   /**